added src build data libs and include dirs, split code from cMchecker.h into seperate files
