+++ libs/gamma.h
