+++ libs/hash.h
