+++ libs/nr3a.h
