+++ libs/ran.h
